requirements.txt
cassandra-driver
pyspark
kafka-python
requests